Initial Branch
Not sure if it works, but basically at the start of run time the code
checks to see where the directory of files is located (it can be
different depending on whether the code is run from a jar or eclipse),
and then loads all files in them. I also would like to have the option
of only loading just enough assets to accurately construct a world, or
specify assets in a file. Currently assets are stored in AssetManager
only by their simple file name (the file, minus the extension).