Fixing values
When a projectile dies and onDeath() is called, if the ability has an
item name set, it will spawn an item entity which will relocate itself
according to regular physics. After which it will assume the idle state.
All the work occurs in AbilityController and ItemController. To set an
item for an ability, simply specify the name of the item entity in the
json for that ability (see test_item.json).

I'm 50-50 on this implementation, maybe need to rework ability-item
system to increase flexibility, but this works.